Refactor code to break monolithic lib source
