+++ examples/ninja/server/rooms/room.js
@@ -1,5 +1,5 @@
 const Room = require('colyseus').Room;
-const ServerLib = require('../../../../server-lib');
+const ServerLib = require('../../../../src/server');
 const g = new ServerLib();
 
 module.exports = class MyRoom extends Room {
@@ -8,27 +8,27 @@ module.exports = class MyRoom extends Room {
     g.setupCharacters('players');
     g.setupResources('trees');
     g.setupBoard(500, 500, '909090');
-    g.createResource('trees', 50, 50);
+    g.createAResource('trees', 50, 50);
   }
 
   onJoin(client) {
-    g.createCharacter('players', client, { x: 200, y: 200 });
+    g.createACharacter('players', client, { x: 200, y: 200 });
   }
 
   onMessage(client, data) {
-    const player = g.getCharacter('players', client);
+    const player = g.getACharacter('players', client);
     const speed = 5;
     const actions = {
       moveUp: () => (player.y -= speed),
       moveDown: () => (player.y += speed),
       moveLeft: () => (player.x -= speed),
       moveRight: () => (player.x += speed),
-      click: () => g.createResource('trees', data.x, data.y),
+      click: () => g.createAResource('trees', data.x, data.y),
     };
     g.handleActions(actions, data);
   }
 
   onLeave(client) {
-    g.deleteCharacter('players', client);
+    g.deleteACharacter('players', client);
   }
 };
