+++ package.json
@@ -22,7 +22,8 @@
   },
   "homepage": "https://github.com/codechangers/io-game-lib#readme",
   "dependencies": {
-    "colyseus.js": "^0.9.1"
+    "colyseus.js": "^0.9.1",
+    "ws": "^7.3.1"
   },
   "devDependencies": {
     "eslint": "^7.10.0",
